Add new modal dialog for sync success
