+++ app/src/main/java/com/tari/android/wallet/ui/screen/home/overview/HomeOverviewModel.kt
@@ -22,5 +22,8 @@ class HomeOverviewModel {
 
         val activeMinersCount: Int? = null,
         val isMining: Boolean = false,
+
+        val showWalletSyncSuccessDialog: Boolean = false,
+        val showWalletRestoreSuccessDialog: Boolean = false,
     )
 }
