+++ app/src/main/res/values/strings.xml
@@ -223,6 +223,9 @@
     <string name="home_empty_state_description">Once you receive some tXTM, you’ll see it here.</string>
     <string name="home_empty_state_button">Start Mining Tari</string>
     <string name="home_tx_list_recent_activity">Recent activity</string>
+    <string name="home_sync_dialog_title">Welcome back!</string>
+    <string name="home_sync_dialog_description">Your wallet has been successfully restored.</string>
+    <string name="home_sync_dialog_button">Get Started!</string>
 
     <!-- tx list -->
     <string name="tx_list_sent_a_payment">%s sent a payment</string>
